Layout, style and QOL
Finished responsively styling the keypad, the display and all individual keys. The arrow keys are properly rotated as well, shadows are being generated and the colors are all as they should be. Responsive design can be applied to most major phone sizes without breaking, with display, keysize and navigation bar dynamically adjusting

Signed-off-by: MArnold-Y <arnold.matthias.y@gmail.com>